thread: Count only "create" time for threads on FreeRTOS
FreeRTOS doesn't have the concept of creating a thread and keeping it
from being scheduled. A thread is created and ready to be scheduled. If
it has greater priority than current one, it will start running. So it's
pointless to count a "creation" and a "start" time for threads.

This patch basically keeps "start" time zeroed and properly count
"create" time - but it should be noted on reports that it's a single
create and start time.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>