freertos: Revamp thread resource collection
While using static allocation avoids the need to manage RAM usage by
threads, there are still some resources associated to threads after they
exit on FreeRTOS. Those resources are collected when the idle thread
runs.

In the benchmark code, however, tens of thousands of threads are created
during a normal run. This makes quite hard to find a time to run the
idle thread: run it every time to collect a few threads, and it takes a
long time to run all the tests; run it once for a longer time, and you
also risk not running it for long enough to collect all resources.

This patch solves that by noting the fact that if a thread is deleted
from *another* thread, the resource collection is done synchronously in
the calling thread - instead of when a thread deletes itself, in which
case only when the idle thread runs the resources are collected.

With that in mind, this patch introduces a `bench_collect_resources`
call, used after tests that create but don't delete the threads, to
perform this removal. To that extent, `bench_thread_exit` now adds the
exiting thread to a list that is collected at `bench_collect_resources`.
As threads can't exit on FreeRTOS, the final line of `bench_thread_exit`
simply takes a semaphore that is never given.

For other systems, the default implementation of
`bench_collect_resources` is empty.

Finally, this approach avoids rethinking all the tests to explicitly
aborting ancillary threads: while this approach would work, some
resource cleanup, like mutexes and the like, would be made harder.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>